Start work on integration tests for those endpoints, fix some 404s
