+++ server/service/hosts.go
@@ -43,6 +43,12 @@ func hostDetailResponseForHost(ctx context.Context, svc fleet.Service, host *fle
 	}, nil
 }
 
+func (svc *Service) FlushSeenHosts(ctx context.Context) error {
+	// No authorization check because this is used only internally.
+	hostIDs := svc.seenHostSet.getAndClearHostIDs()
+	return svc.ds.MarkHostsSeen(ctx, hostIDs, svc.clock.Now())
+}
+
 ////////////////////////////////////////////////////////////////////////////////
 // List Hosts
 ////////////////////////////////////////////////////////////////////////////////
